[CSS] Add :global() and :local() pseudo-classes
This commit adds `:global(selector)` and `:local(selector)` pseudo-class
support. The former is used by CSS Modules, Less and Astro framework.
The ladder is used by CSS Modules only, but is added for completeness.

It is no default CSS feature, but appears to be specific enough to not cause
any conflicts.

Adding it to core CSS would avoid the need to extend CSS for Astro, just
for this single reason.