[HTML] Add variable for custom tag start character
This commit adds a `tag_name_start` variable to represent a custom tag's first
character across syntax definitions instead of hard-coding `[A-Za-z]`.

It enables template syntaxes to override such tags' names easier, if they begin
with a template interpolation.

Example:

    <{{tag_name}} ...>

To scope `{{tag_name}}` entity.name.tag, a template syntax has needed to extend
the `tag-other` context before this commit with something like:

    tag-other:
      - meta_prepend: true
      - match: </?(?={{)
        scope: punctuation.definition.tag.begin.html
        push:
          - tag-other-content
          - tag-other-name

even though `{{ ... }}` tags are already part of the `prototype` context.

With this commit it is enough to override the `tag_name_start` to add the
template begin punctuation in order to achieve the same outcome.

    variables:
      tag_name_start: (?:[A-Za-z]|{{)