[Batch File] Fix empty comments exceeding to next line
This commit fixes "label comment" related patterns, to only scope leading
colons `punctuation.definition.comment`.

Label comments make use of invalid labels being ignored.
A label is invalid, if the first character after leading colon is a "metachar".

A label starting with `::` is already an invalid label and thus can be followed
by any (meta) char to still be a "valid" comment.

The wrong pattern was the real reason for the issue addressed by PR #3992.