[Batch File] Fix line continuation
This commit applies the following rule:

If `^` appears at the end of a line, the linefeed is stripped and the next
character is escaped or uses its special meaning, if any.
If the next character is also a linefeed, it is treated literal.

REM command seems to be working special by line continuation working only
directly after the first token.