[Batch File] Fix labels
This commit starts implementing tokenization according to descriptions of
https://stackoverflow.com/questions/4094699/how-does-the-windows-command-interpreter-cmd-exe-parse-scripts/4095133#4095133
for lables, which are also used as sort of line comments.

It also highlights the odd parts, which includes not all valid looking labels
being able to targeted in GOTO and CALL statements due to different
tokenization rules being applied by dosbatch interpreter.