[JavaScript] Fix ternary operator regression
This commit fixes a regression, introduced by #3986, which breaks syntax
highlighting in objects, as comma is no longer consumed as key-value-pair
separator.

It's not the ternary expression, which allows commas, but the for statement,
which allows multiple expressions, separated by comma after `in` keyword.

    for (var in <expr>, <expr>, ...)

    expr: any valid expression, maybe ternary, maybe function-call

The change to ternary expressions is reverted and for-condition context is
modified instead to allow multiple expressions after `in` keyword.

Also scope `;` invalid if there are too many.