[ShellScript] Support interpolation in numbers
This commit add support for variable expansion in number to support
some examples illustrated in pure bash bible, which convert colors
to RGB.

        ((r=16#${_:0:2},g=16#${_:2:2},b=16#${_:4:2}))

Illegal highlighting is dropped, but that's probably even more reliable,
especially in scenarios of unknown delimiter or terminator characters.

TODO: ZSH floating points are not covered, yet.