[C] Add support for new preprocessor directives
This adds support for the new #elifdef and #elifndef preprocessor
directives that are a combination of #else with #ifdef or #ifndef.

C++, Objective-C, and Objective-C++ are also updated because they
derive from the C syntax family as well.

This has already made it into recent C and C++ compilers and is
part of the C23 and C++23 standards.