[ShellScript] Fix built-in overrides
This commit reorganizes statement contexts for built-in commands to be
augment-able by local function definitions, while keeping higher priority
of reserved words.

Reserved words, like `if`, `case`, ... can not be used as identifier for
local function definitions, but all other built-in commands can be.

As a result `def-anonymous` and `def-function` contexts can be merged.

`cmd-coproc` is re-defined as kind of function definition under `def-coproc`.

The token `time` is moved to `cmd-control` context
and scoped `keyword.control` as it is a reserved word.