[C] Add support for binary file inclusion preprocessor directive
This adds support for the new #embed preprocessor directive that can
be used to directly include binary files into program variables without
the use of external code generators.

C++, Objective-C, and Objective-C++ are also updated because they derive
from the C syntax family as well.

This has already made it into recent C and C++ compilers and is also
part of the C23 standard.