[HTML] Restrict main HTML snippet
This commit disables `html` snippet in ...

- comments
- literal or interpolated strings
- embedded or interpolated code blocks (from template languages)
- HTML tags
- Markdown files (but not in fenced code blocks)

Using `- text.html <scope>` in order to only exclude scopes belonging to
possibly embedded HTML code block in question.

This is required to enable snippets if HTML is embedded in `source` code,

   source.astro text.html

but disable it in embedded CSS/JS or any other source code.

   text.html source.css