[C] Add support for decimal floating-point types
This was introduced in C23 and adds support for the new type
keywords, the new suffixes for the new types, and the new
length modifiers for printf and scanf.

Furthermore, the existing matching rules for floats was changed
to remove the possibility of '1f' being classified as a floating
point literal. This is actually considered to be an integral
literal with an invalid suffix instead of a floating point literal
by all compilers that I know of and such produces compiler errors
if such code is compiled.