[C] Make function parameter names optional (#4093)
Function parameter names are now optional in C and C++. Previously
the syntax highlighting rules would assume that the last identifier
found in a parameter was the parameter name but this was making it
impossible to properly highlight parameters without a parameter name.

The only solution I could find was to include the specific types and
modifiers before the more general match for the parameter name. With
this, the types and modifiers are no longer mistaken for a parameter
name.