[C] Apply same digit separators for C/C++ (#4101)
C23 added this feature previously available in C++. This should
close #3972 as well as both C and C++ now have syntax support for
the ' digit separator.

This was implemented by just copying the digit variables from the
C++ syntax file as it is the same construct in the end. The existing
tests were updated to fix the part of the existing tests were broken
by this new feature.

Furthermore, some tests from C++ were adapted for C to add better
testing for the new digit seperator. I also included tests for the
newer base 2 integer contants. The C file did not have any tests for
that yet.

Co-authored-by: deathaxe <deathaxe82@googlemail.com>