[HTML] Fix various conflict marker issues
This commit explicitly includes `merge-conflict-markers` in front of
each tag end punctuation (`>` or `/>`) to ensure conflict markers
taking precedence.

Including that context in `tag-attributes` is not sufficient as it is
included after tag end punctuation patterns.

Including that context in `prototype` is not sufficient as as it is
excluded from tag attribute contexts for templating languages to inject
template tags only within attribute names or values using `prototype`.

Syntax tests are reorganized in a dedicated test file.