[ZSH] Fix parameter identifier pattern (#4245)
This commit fixes parameter/variable identifier patterns to comply with
section 15.1 of ZSH language specification.

    A parameter has a name, a value, and a number of attributes. A name may be
    any sequence of alphanumeric characters and underscores, or the single
    characters ‘*’, ‘@’, ‘#’, ‘?’, ‘-’, ‘$’, or ‘!’. A parameter whose name
    begins with an alphanumeric or underscore is also referred to as a
    variable.

    Variables consist of alphanumeric characters and thus can even start with
    numeric characters.

Bash's `identifier` and `variable` patterns are tweaked in order to be able to
implement different ZSH behavior by overriding a single variable, only.

Various tests are added to verify this pattern not only in variable assignment
statements but also parameter expansions.