[ZSH] Support more sophisticated function definition names
This commit is a follow-up of #4262 to add support for more sorts of quoting
and interpolation within function definition names beyond scope of issue #4244.

It may not catch any insane edge case, but hopefully the most likely ones.
Anything beyond would require branching to distinguish normal command calls
from bare function definitions. It is not judged worth it at this point.