[Various] Scope `in` as `keyword.control` again (#4271)
This is a follow-up to 1e76a91, which introduced this new unified
`keyword.operator.iteration.in` scope. However, this `in` is a keyword
that is used in loop statements and not an operator and should thus be
scoped similar to `keyword.control.loop.for`. I opted to simply replace
`for` with `in` in the above as it allows for sufficient context of the
(loop-related control flow) keyword while not binding it to the `for`
like it was used in some instances before.