[C] Recognise types ending in _t
This is one place where VS Code does better with the default syntax,
because it recognises the majority of custom types as-is without any
sort of structural type analysis. (This also works better than analysis,
as that necessarily requires the LSP to be able to perform analysis,
which isn't always the case).

https://github.com/microsoft/vscode/blob/3aef8be5e46aa7e97e8b0ed115f6147cd0b24634/extensions/cpp/syntaxes/c.tmLanguage.json#L195-L198