[Markdown] Fix typing ``` in the middle of a document (#4363)
Typing ``` was inserting ```` (except when typing at EOF), because the initial
paired `` is not recognized by the syntax definition as a code block, and
therefore the second backtick was inserting, rather than skipping over the
paired character.

Additionally deleted the bindings that would delete ** and __ sequences with one
keypress, because they're not auto-paired, so it doesn't make any sense to do
this.