[Python] Unify block begin punctuation
This commit scopes all colons `punctuation.section.block.begin`, which start
an indented block of code, regardless their actual context.

The point is colon indicating a start of a code block, like an opening brace
in a C-style language. We just don't scope the block `meta.block` as we don't
want or can't reliably count indentation to pop it off, correctly.